Copy the names of the waste loads and the monitoring stations for use in BMI
Just as the names of the substances we need to copy the names of the waste loads and the monitoring stations, so that they are know in a convenient way to the BMI routines.
Note: we assume the _first_ segment in a dump/monitoring area is representative. Actually we should probably exclude monitoring areas with more than one segment.