UNST-7671 new approach to setting nc_types in out_quan_conf_his
Changed the way the nc_type of variables in the out_quan_conf_his is defined.

The list of all possible output variables contains the netcdf precision with which each variable should be written. We want this to be configurable through the mdu for most (but not all!) variables, but the list of possible output variables is defined before the mdu is read, which is annoying.

For this reason, the variable configs in the output list now contain the "id_nc_type", which can indicate any of the existing netcdf types (int, float, double, etc.), but also "undefined". In the calls to definencvar, these ids are converted to actual nc_type's, with "undefined" being converted to the nc_type defined by the mdu (either float or double).