UNST-7889 move updateBalance to finalize_single_timestep
Moved the call to updateBalance (which reduces the integrated water balance variables across the processes) from finalize_user_timestep to finalize_single_timestep. This way, the values in voltot are accurate in every time step, and statistics in time can still be performed on them.

Also removed the exception for UNC_LOC_GLOBAL in reduce_statistical_output.