UNST-7786 refactor cross-sections code
Refactored the code that ultimately fills the observation cross-sections array crs.

Made it so that the integrals over the cross-sections are calculated and mpi-reduced in every model time step, so that statistical output always has accurate values to work with.

Verified locally with test case e02\f014\c001. On to the test bench!