Rename "substance_index"
As the substance index may at some point be used in a more general sense, I have renamed it to quantity index. This is in line with the earlier change of substance name to quantity name. Note that in some places in the code we do need to use "substance" as they refer to the actual list of substances in the calculation.