Introduce an extra exchange parameter - previous volume
The numerical difficulties I encountered turned out to be due to a phase shift in the volumes that are transferred between D-Flow FM and DELWAQ. By introducing an extra exchange parameter, this problem has been solved. Or at least, the numerical problems have gone. The next step will be to check the continuity, but the first results are encouraging.