DELWAQ-1199: Apply the absolute BLOOM time step to the process management routines
As the actual "fractional" time stepping for BLOOM is done in the management routines (calculate_processes and its D-Flow FM equivalent), these routines need to be aware of the change as well. Be careful that a variable time step may upset the count. As the derivatives are applied at the first time step of the group, not proportionally, a "local" estimate of the number of steps that now fit into the interval should do.
Also correct a few mistakes in the process routines that are involved.