PART-119: Make sure all volumes are positive
During the examination of the test cases I ran into issues with a division by zero because the volume of a grid cell in a z-layers model was zero. The grid cell was not a bottom cell. This led to a unification of the volume corrections. sigma-layers and z-layers are handled in the same way.